move collocation folder directly into pages folder
