+++ platform/cypress/e2e/collocation/device_status_table.cy.js
@@ -1,6 +1,6 @@
 describe('DataTable', () => {
   beforeEach(() => {
-    cy.visit('/analytics/collocation/collocate');
+    cy.visit('/collocation/collocate');
   });
 
   it('displays the table headers', () => {
