+++ platform/src/pages/account/login/index.jsx
@@ -50,7 +50,7 @@ const UserLogin = () => {
             dispatch(setUserInfo(response.users[0]));
             dispatch(setSuccess(true));
             setLoading(false);
-            router.push('/analytics/collocation/overview');
+            router.push('/collocation/overview');
           })
           .catch((error) => {
             console.error(`Error fetching user details: ${error}`);
@@ -108,7 +108,8 @@ const UserLogin = () => {
               <button
                 data-testid='login-btn'
                 className='mt-6 btn bg-blue-900 rounded-none w-full text-sm outline-none border-none hover:bg-blue-950'
-                type='submit'>
+                type='submit'
+              >
                 {loading ? <Spinner data-testid='spinner' width={25} height={25} /> : 'Login'}
               </button>
             </div>
