--- platform/src/pages/analytics/airqlouds.jsx
@@ -1,10 +1,0 @@
-import Layout from '@/components/Layout';
-import withAuth from '@/core/utils/protectedRoute';
-
-const AirQlouds = () => (
-  <Layout>
-    <div></div>
-  </Layout>
-);
-
-export default withAuth(AirQlouds);
