+++ platform/src/pages/collocation/collocate.jsx
@@ -78,7 +78,8 @@ const Collocate = () => {
               <Button
                 className={
                   'bg-white text-black-600 border border-black-600 opacity-30 hover:cursor-not-allowed font-medium text-sm'
-                }>
+                }
+              >
                 <div className='mr-[10px]'>
                   <UploadIcon />
                 </div>
@@ -89,7 +90,8 @@ const Collocate = () => {
                 className={
                   'rounded-none text-white bg-blue-900 border border-blue-900 hover:bg-dark-blue hover:border-dark-blue font-medium text-sm'
                 }
-                path='/analytics/collocation/add_monitor'>
+                path='/collocation/add_monitor'
+              >
                 <div className='mr-[10px]'>
                   <BoxedAddIcon />
                 </div>
