+++ platform/src/pages/collocation/reports/monitor_report/[device].jsx
